pyproject.toml
[tool.poetry]
name = "rehabilitation-bot"
version = "0.1.0"
description = "Telegram bot of the rehabilitation community at the Konstantin Khabensky Foundation"
authors = [
    "Юлий Митряшкин", "Яна Денисова",
    "Никита Викторов", "Алексей Соловьев", "Марк Марузов"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.1"
poetry = "^1.4.2"
pydantic = "^1.10.7"
python-telegram-bot = "^20.2"
emoji = "^2.4.0"
thefuzz = {extras = ["speedup"], version = "^0.19.0"}
sqlalchemy = "^2.0.15"
alembic = "^1.11.1"
asyncpg = "^0.27.0"
async-lru = "^2.0.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
python-dotenv = "0.20.0"
pre-commit = "^3.2.2"
autopep8 = "2.0.2"
add-trailing-comma = "2.4.0"
mypy = "1.2.0"
flake8-pyproject = "^1.2.3"
aiosqlite = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
default_section = "THIRDPARTY"
sections = ["STDLIB", "FIRSTPARTY", "THIRDPARTY", "LOCALFOLDER"]
known_local_folder = ["src", "bot", "core"]
lines_after_imports = "2"
skip = ["tests", ".venv", "venv"]
line_length = "79"
multi_line_output = "3"
include_trailing_comma = "true"
use_parentheses = "true"

[tool.flake8]
max-line-length = "79"
paths = ["./bot", "./core", "./src"]
max-complexity = "10"
exclude = ["tests", ".venv", "venv"]
per-file-ignores = [
    "__init__.py:F401",
]
